+++ internal/graph/sync/extractors/flux_helmrelease_test.go
@@ -289,7 +289,7 @@ func TestFluxHelmReleaseExtractor_ConfidenceScoring(t *testing.T) {
 			name: "all evidence present",
 			setupLookup: func(m *mockResourceLookup, hrTimestamp int64) {
 				// Resource created 5 seconds after HelmRelease
-				deploymentTime := hrTimestamp + (5 * int64(time.Second.Nanoseconds()))
+				deploymentTime := hrTimestamp + (5 * time.Second.Nanoseconds())
 				m.AddResource("deploy-uid", "Deployment", "default", "frontend", deploymentTime)
 
 				// Add reconcile event
@@ -308,7 +308,7 @@ func TestFluxHelmReleaseExtractor_ConfidenceScoring(t *testing.T) {
 			name: "label match only",
 			setupLookup: func(m *mockResourceLookup, hrTimestamp int64) {
 				// Resource created long ago (no temporal evidence)
-				oldTime := hrTimestamp - (60 * int64(time.Second.Nanoseconds()))
+				oldTime := hrTimestamp - (60 * time.Second.Nanoseconds())
 				m.AddResource("deploy-uid", "Deployment", "default", "frontend", oldTime)
 			},
 			expectedConfidence: struct{ min, max float64 }{min: 0.4, max: 0.6},
@@ -317,7 +317,7 @@ func TestFluxHelmReleaseExtractor_ConfidenceScoring(t *testing.T) {
 			name: "no evidence",
 			setupLookup: func(m *mockResourceLookup, hrTimestamp int64) {
 				// Resource with different name, created long ago
-				oldTime := hrTimestamp - (60 * int64(time.Second.Nanoseconds()))
+				oldTime := hrTimestamp - (60 * time.Second.Nanoseconds())
 				m.AddResource("deploy-uid", "Deployment", "default", "backend", oldTime)
 			},
 			expectedConfidence: struct{ min, max float64 }{min: 0.0, max: 0.2},
@@ -383,7 +383,7 @@ func TestFluxHelmReleaseExtractor_TargetNamespace(t *testing.T) {
 
 	lookup := newMockResourceLookup()
 	hrTimestamp := time.Now().UnixNano()
-	deployTime := hrTimestamp + (3 * int64(time.Second.Nanoseconds()))
+	deployTime := hrTimestamp + (3 * time.Second.Nanoseconds())
 
 	// Add resource in production namespace
 	lookup.AddResource("deploy-uid", "Deployment", "production", "frontend", deployTime)
